Use j2cli to interpolate molecule platform list
This replaces the existing sed-based interpolation with Jinja2
templates, using j2cli to compile them. This removes some bash logic embedded
in JenkinsfilePS, moving it to the template and the configuration
file that feeds it. The templates require the PLATFORMS environment
variable to be compiled, but not the ROLE_NAME environment variable.